+++ docs/changelog.md
@@ -4,6 +4,7 @@
 1. fix editor's configuration could affect each other bug.
 2. add preview to picture upload
 3. add type check for initialize config
+4. fix content sync bug when switch to source code view
 
 0.2.8